# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MenuClick API",
    "version": "0.1.0",
    "description": "API do MenuClick — cardápio digital, QR code e delivery.\n\n## Duas superfícies\n\nA API atende dois públicos, e a diferença é o que mais importa ao usá-la:\n\n- **Quem escaneia o QR code** lê o cardápio e faz pedido, sem conta e sem token.\n- **O restaurante** faz todo o resto, com `Authorization: Bearer <token>`.\n\n**Toda rota exige sessão por padrão.** Ser pública é declaração explícita,\ne as poucas que são estão marcadas como tal aqui.\n\n## Convenções\n\n- Listagens respondem um envelope `{ data, limit, offset, total }`, nunca um array cru.\n  `limit` vai de 1 a 100 (default 20); fora da faixa é **400**, não ajuste silencioso.\n- Erro tem sempre a forma `{ statusCode, error, message }`.\n- Valores monetários são **inteiros em centavos**. Nunca float.\n- Nada é apagado: `DELETE` marca o registro como removido, e ele passa a se\n  comportar como se nunca tivesse existido (404 em tudo, fora das listagens).\n- Há limite de requisições por IP: 100/min no geral e **5/min no login**.\n  Estourar responde **429**."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "O token devolvido por `POST /auth/login`. É opaco: não carrega informação, e o servidor guarda só o hash dele."
      }
    },
...
  },
  "paths": {
...
    "/restaurants/{restaurantId}/orders/summary": {
      "get": {
        "operationId": "getOrdersSummary",
        "summary": "Resumo do período para o painel",
        "tags": [
          "Pedidos"
        ],
        "description": "Contadores por status, faturamento e ticket médio. Faturamento é o que o restaurante **aceitou vender**: de `confirmed` em diante, sem `pending` (ainda não é venda) nem `cancelled` (deixou de ser) — contar só `completed` mostraria quase zero no pico do almoço. Aceita o mesmo recorte da listagem (`period` **ou** `from`/`to`), resolvido no fuso do restaurante, e devolve em `period` os instantes que usou.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "yesterday",
                "last7days",
                "thisMonth"
              ]
            },
            "in": "query",
            "name": "period",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "date"
            },
            "in": "query",
            "name": "from",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "date"
            },
            "in": "query",
            "name": "to",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "restaurantId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "period": {
                      "type": "object",
                      "properties": {
                        "from": {
                          "type": "string"
                        },
                        "to": {
                          "type": "string"
                        }
                      }
                    },
                    "counts": {
                      "type": "object",
                      "properties": {
                        "pending": {
                          "type": "integer"
                        },
                        "confirmed": {
                          "type": "integer"
                        },
                        "preparing": {
                          "type": "integer"
                        },
                        "ready_for_pickup": {
                          "type": "integer"
                        },
                        "out_for_delivery": {
                          "type": "integer"
                        },
                        "completed": {
                          "type": "integer"
                        },
                        "cancelled": {
                          "type": "integer"
                        }
                      }
                    },
                    "revenueInCents": {
                      "type": "integer"
                    },
                    "revenueOrderCount": {
                      "type": "integer"
                    },
                    "averageTicketInCents": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/orders/{orderId}": {
      "get": {
        "operationId": "getTrackedOrder",
        "summary": "Lê o pedido pelo token de acompanhamento",
        "tags": [
          "Pedidos"
        ],
        "description": "A leitura HTTP do mesmo pedido que o WebSocket acompanha, e com os **itens** — que o canal não transmite. Serve para a tela sobreviver a um reload e como alternativa quando o socket não conecta. Autoriza pelo `token` devolvido na criação; pedido de salão não recebe token e não é legível por aqui. Não devolve os dados do cliente nem do restaurante.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "query",
            "name": "token",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "orderId",
            "required": true
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "totalInCents": {
                      "type": "integer"
                    },
                    "deliveryFeeInCents": {
                      "type": "integer",
                      "nullable": true
                    },
                    "paymentMethod": {
                      "type": "string"
                    },
                    "changeForInCents": {
                      "type": "integer"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "productId": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "priceInCents": {
                            "type": "integer"
                          },
                          "unitPriceInCents": {
                            "type": "integer"
                          },
                          "quantity": {
                            "type": "integer"
                          },
                          "options": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "groupName": {
                                  "type": "string"
                                },
                                "name": {
                                  "type": "string"
                                },
                                "priceInCents": {
                                  "type": "integer"
                                },
                                "quantity": {
                                  "type": "integer"
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "deliveryAddress": {
                      "type": "object",
                      "properties": {
                        "street": {
                          "type": "string",
                          "minLength": 1
                        },
                        "number": {
                          "type": "string",
                          "minLength": 1
                        },
                        "neighborhood": {
                          "type": "string",
                          "minLength": 1
                        },
                        "city": {
                          "type": "string",
                          "minLength": 1
                        },
                        "state": {
                          "type": "string",
                          "minLength": 1
                        },
                        "zipCode": {
                          "type": "string",
                          "minLength": 1
                        }
                      }
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "Desenvolvimento"
    }
  ],
...
}